+++ lib/api_trakt/exceptions.py
@@ -47,3 +47,7 @@ class TraktLockedUserAccount(TraktException):
 
 class TraktInvalidGrant(TraktException):
     pass
+
+
+class TraktFreemiumLimit(TraktException):
+    pass